Created disclosure-lib sub-project
Updates build files to support new library.  Removed external library
proven-disclosure.
