--- proven-disclosure/README.md
@@ -1,2 +1,0 @@
-# proven-disclosure
-Proven disclosure library, supporting message disclosure to a proven-cluster
