--- proven-disclosure/proven-disclosure.gradle
@@ -1,67 +1,0 @@
- apply plugin: 'java'
- apply plugin: 'eclipse'
- apply plugin: 'maven-publish'
- 
-  
- sourceCompatibility = 1.8
- version = '0.1'
-  
-jar {
-
- }
- 
- 
- task javadocJar(type: Jar) {
-   from javadoc
-   classifier 'javadoc'
- }
- 
-task sourcesJar(type: Jar) {
-   from sourceSets.main.allJava
-   classifier 'sources'
-}
-
-task allInOneJar(type: Jar) {
-    manifest {
-        attributes 'Main-Class': 'gov.pnnl.proven.message.ProvenMessage'
-    }
-    from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
-    with jar
-    classifier 'all-in-one'
-}
-    
- publishing {
-    publications {
-        mavenJava(MavenPublication) {
-        	groupId 'gov.pnnl.proven'
-        	artifactId 'proven-message'
-            from components.java
-            artifact sourcesJar
-            artifact allInOneJar
-            artifact javadocJar
-        }
-	}
-}	
-
- repositories {
-	 mavenCentral()
-	 mavenLocal()
- }
- 
- dependencies {
- 	compile group: 'javax.json', name: 'javax.json-api', version: '1.1.3'
- 	compile group: 'org.glassfish', name: 'javax.json', version: '1.1.3'
- }
- 
- test {
-	 systemProperties 'property': 'value'
- }
- 
- uploadArchives {
-	 repositories {
-		flatDir {
-			dirs 'repos'
-		}
-	 }
- }
-
