--- proven-disclosure/src/main/java/gov/pnnl/proven/client/lib/disclosure/examples/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Provides example uses of proven-disclosure library 
- * 
- * @author d3j766
- *
- */
-package gov.pnnl.proven.client.lib.disclosure.examples;
