--- proven-disclosure/src/main/java/gov/pnnl/proven/client/lib/disclosure/request/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Provides a set of requests provided with this library.
- * 
- * @author d3j766
- *
- */
-package gov.pnnl.proven.client.lib.disclosure.request;
