+++ proven-member/disclosure-lib/disclosure-lib.gradle
@@ -1,0 +1,55 @@
+ apply plugin: 'java'
+ apply plugin: 'eclipse'
+ apply plugin: 'maven-publish'
+ 
+  
+ sourceCompatibility = 1.8
+ version = '0.1'
+  
+ 
+ task javadocJar(type: Jar) {
+   from javadoc
+   classifier 'javadoc'
+ }
+ 
+task sourcesJar(type: Jar) {
+   from sourceSets.main.allJava
+   classifier 'sources'
+}
+
+task allInOneJar(type: Jar) {
+    from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
+    classifier 'all-in-one'
+    with jar
+}
+    
+ publishing {
+    publications {
+        mavenJava(MavenPublication) {
+        	groupId 'gov.pnnl.proven'
+        	artifactId 'proven-message'
+            from components.java
+            artifact sourcesJar
+            artifact allInOneJar
+            artifact javadocJar
+        }
+	}
+}	
+
+ repositories {
+	 mavenCentral()
+	 mavenLocal()
+ }
+ 
+dependencies { 	      
+ 	compile group: 'javax.json', name: 'javax.json-api', version: '1.1.3'
+ 	compile group: 'org.glassfish', name: 'javax.json', version: '1.1.3'
+}
+
+test { systemProperties 'property': 'value'   }
+
+uploadArchives {
+	repositories { flatDir { dirs 'repos'
+		} }
+}
+
