+++ proven-member/disclosure-lib/src/main/java/gov/pnnl/proven/cluster/lib/disclosure/ClientResponseMap.java
@@ -37,78 +37,39 @@
  * PACIFIC NORTHWEST NATIONAL LABORATORY operated by BATTELLE for the 
  * UNITED STATES DEPARTMENT OF ENERGY under Contract DE-AC05-76RL01830
  ******************************************************************************/
-package gov.pnnl.proven.module.disclosure;
+package gov.pnnl.proven.cluster.lib.disclosure;
 
-import java.util.List;
-import java.util.Set;
-
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
-import com.hazelcast.core.Member;
-import com.hazelcast.ringbuffer.Ringbuffer;
 
-import fish.payara.micro.PayaraMicro;
-import gov.pnnl.proven.cluster.lib.disclosure.ClientDisclosureMap;
-import gov.pnnl.proven.cluster.lib.disclosure.ClientResponseMap;
-import gov.pnnl.proven.cluster.lib.disclosure.ProxyRequest;
-import gov.pnnl.proven.cluster.lib.module.ProvenModule;
+/**
+ * Provides the name of a Proven Cluster's IMap which contains REST service
+ * URL's that can be used as targets for a Server Sent Event (SSE) subscription,
+ * allowing Proven Clients to listen for response data from requests previously
+ * submitted to the cluster. The SSE event contains {@code ClientResponseData}
+ * in its data field.
+ * 
+ * The IMap Key is a String containing the REST URL and is defined as (SESSION
+ * must be replaced with the {@code ClientProxyRequest} session value):
+ * 
+ * GET http://<host>:<port>/<context_root>/SESSION/responses
+ * 
+ * The IMap Value is a boolean. True indicates the service is available for use.
+ * False indicates the service is not available for use.
+ * 
+ * @author d3j766
+ * 
+ * @see ClientResponseData
+ * @since
+ *
+ */
+public class ClientResponseMap {
 
-@ApplicationScoped
-public class DisclosureModule extends ProvenModule {
+	static Logger log = LoggerFactory.getLogger(ClientResponseMap.class);
+	public static final String CLIENT_RESPONSE_MAP_NAME = "client.response_map";
 
-	private static Logger log = LoggerFactory.getLogger(DisclosureModule.class);
-
-	public static final String DISCLOSURE_BUFFER = "disclosure.buffer";
-
-	public static final String HOST_TAG = "<HOST>";
-	public static final String PORT_TAG = "<PORT>";
-	public static final String SESSION_TAG = "SESSION";
-	public static final String RESPONSE_URL_TEMPLATE = "http://" + HOST_TAG + ":" + PORT_TAG + "/disclosure/"
-			+ SESSION_TAG + "/responses";
-
-	@Inject
-	private HazelcastInstance hzInstance;
-
-	Ringbuffer<ProxyRequest<?, ?>> disclosureBuffer;
-
-	IMap<String, Boolean> clientDisclosureMap;
-
-	IMap<String, Boolean> clientResponseMap;
-
-	@PostConstruct
-	public void init() {
-
-		// TODO This should be part of the member registry when a
-		// DisclosureBuffer
-		// reports itself as part of its construction. Placed here for now to
-		// support moving message-lib processing to cluster.
-		disclosureBuffer = hzInstance.getRingbuffer(DISCLOSURE_BUFFER);
-		clientDisclosureMap = hzInstance.getMap(new ClientDisclosureMap().getDisclosureMapName());
-		clientDisclosureMap.put(DISCLOSURE_BUFFER, true);
-		clientResponseMap = hzInstance.getMap(new ClientResponseMap().getResponseMapName());
-		String responseUrl = buildResponseUrl();
-		clientResponseMap.put(responseUrl, true);
-		//testPipeline();
-		log.debug("DisclossureModule constructed");
-	}
-	
-	private String buildResponseUrl() {
-		String ret;
-		Member member = hzInstance.getCluster().getLocalMember();
-		String host = member.getAddress().getHost();
-		String port = PayaraMicro.getInstance().getRuntime().getLocalDescriptor().getHttpPorts().get(0).toString();
-		ret = RESPONSE_URL_TEMPLATE.replace(HOST_TAG, host);
-		ret = ret.replace(PORT_TAG, port);
-		return ret;
-	}
-	
-	private void testPipeline() {
-		new TestPipeline().submit();
-	}
+	public String getResponseMapName() {
+		return CLIENT_RESPONSE_MAP_NAME;
+	};
 
 }
