+++ proven-member/disclosure-lib/src/main/java/gov/pnnl/proven/cluster/lib/disclosure/DisclosureDomain.java
@@ -37,78 +37,60 @@
  * PACIFIC NORTHWEST NATIONAL LABORATORY operated by BATTELLE for the 
  * UNITED STATES DEPARTMENT OF ENERGY under Contract DE-AC05-76RL01830
  ******************************************************************************/
-package gov.pnnl.proven.module.disclosure;
+package gov.pnnl.proven.cluster.lib.disclosure;
 
-import java.util.List;
-import java.util.Set;
-
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
+import java.io.Serializable;
+import org.apache.commons.validator.routines.DomainValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
-import com.hazelcast.core.Member;
-import com.hazelcast.ringbuffer.Ringbuffer;
 
-import fish.payara.micro.PayaraMicro;
-import gov.pnnl.proven.cluster.lib.disclosure.ClientDisclosureMap;
-import gov.pnnl.proven.cluster.lib.disclosure.ClientResponseMap;
-import gov.pnnl.proven.cluster.lib.disclosure.ProxyRequest;
-import gov.pnnl.proven.cluster.lib.module.ProvenModule;
+import gov.pnnl.proven.cluster.lib.disclosure.exception.InvalidDisclosureDomainException;
 
-@ApplicationScoped
-public class DisclosureModule extends ProvenModule {
+/**
+ * Information disclosed by external clients to a Proven Cluster are organized
+ * under domain names. A valid domain name used by Proven must have a recognized
+ * top-level domain as defined in {@code DomainValidator}
+ * 
+ * @author d3j766
+ *
+ * @see DomainValidator
+ * @since
+ * 
+ */
+public class DisclosureDomain implements DomainProvider, Serializable {
 
-	private static Logger log = LoggerFactory.getLogger(DisclosureModule.class);
+	private static final long serialVersionUID = 1L;
+	static Logger log = LoggerFactory.getLogger(DisclosureDomain.class);
 
-	public static final String DISCLOSURE_BUFFER = "disclosure.buffer";
+	/**
+	 * Domain name.
+	 */
+	String domain;
 
-	public static final String HOST_TAG = "<HOST>";
-	public static final String PORT_TAG = "<PORT>";
-	public static final String SESSION_TAG = "SESSION";
-	public static final String RESPONSE_URL_TEMPLATE = "http://" + HOST_TAG + ":" + PORT_TAG + "/disclosure/"
-			+ SESSION_TAG + "/responses";
+	public DisclosureDomain(String domain) throws InvalidDisclosureDomainException {
 
-	@Inject
-	private HazelcastInstance hzInstance;
-
-	Ringbuffer<ProxyRequest<?, ?>> disclosureBuffer;
-
-	IMap<String, Boolean> clientDisclosureMap;
-
-	IMap<String, Boolean> clientResponseMap;
-
-	@PostConstruct
-	public void init() {
-
-		// TODO This should be part of the member registry when a
-		// DisclosureBuffer
-		// reports itself as part of its construction. Placed here for now to
-		// support moving message-lib processing to cluster.
-		disclosureBuffer = hzInstance.getRingbuffer(DISCLOSURE_BUFFER);
-		clientDisclosureMap = hzInstance.getMap(new ClientDisclosureMap().getDisclosureMapName());
-		clientDisclosureMap.put(DISCLOSURE_BUFFER, true);
-		clientResponseMap = hzInstance.getMap(new ClientResponseMap().getResponseMapName());
-		String responseUrl = buildResponseUrl();
-		clientResponseMap.put(responseUrl, true);
-		//testPipeline();
-		log.debug("DisclossureModule constructed");
+		// Get a DomainValidator
+		DomainValidator validator = DomainValidator.getInstance();
+		if (!validator.isValid(domain)) {
+			throw new InvalidDisclosureDomainException();
+		}
 	}
-	
-	private String buildResponseUrl() {
-		String ret;
-		Member member = hzInstance.getCluster().getLocalMember();
-		String host = member.getAddress().getHost();
-		String port = PayaraMicro.getInstance().getRuntime().getLocalDescriptor().getHttpPorts().get(0).toString();
-		ret = RESPONSE_URL_TEMPLATE.replace(HOST_TAG, host);
-		ret = ret.replace(PORT_TAG, port);
+
+	@Override
+	public DisclosureDomain getDomain() {
+		return this;
+	}
+
+	public String getReverseDomain() {
+		String ret = "";
+		String[] parts = domain.split(".");
+		int start = 0;
+		int end = parts.length;
+		while (start < end) {
+			ret += parts[start];
+			start++;
+		}
 		return ret;
-	}
-	
-	private void testPipeline() {
-		new TestPipeline().submit();
 	}
 
 }
