+++ proven-member/disclosure-lib/src/main/java/gov/pnnl/proven/cluster/lib/disclosure/examples/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Provides example uses of proven-gov.pnnl.proven.cluster.lib.disclosure library 
+ * 
+ * @author d3j766
+ *
+ */
+package gov.pnnl.proven.cluster.lib.disclosure.examples;
