+++ proven-member/disclosure-lib/src/main/java/gov/pnnl/proven/cluster/lib/disclosure/request/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Provides a set of requests provided with this library.
+ * 
+ * @author d3j766
+ *
+ */
+package gov.pnnl.proven.cluster.lib.disclosure.request;
