+++ proven-member/disclosure-lib/src/main/resources/log4j2.xml
