+++ settings.gradle
@@ -3,6 +3,7 @@ include "proven-member:member-lib"
 include "proven-member:exchange-lib"
 include "proven-member:module-lib"
 include "proven-member:message-lib"
+include "proven-member:disclosure-lib"
 include "proven-member:stream-lib"
 include "proven-member:hybrid-lib"
 include "proven-member:stream-module"
@@ -10,7 +11,6 @@ include "proven-member:disclosure-module"
 include "proven-member:hybrid-module"
 include "proven-member:member-module"
 include "proven-client"
-include "proven-disclosure"
 include "proven-docker"
